requirements.txt
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
export==0.1.2
Flask==1.1.2
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.5
pandas==1.2.1
Pillow==8.1.0
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.5
requests==2.25.1
six==1.15.0
toml==0.10.2
txt==2020.11.12
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1